refactor(serialization): Separate serializer into components
`serializer.c` was the combination of all the serialization functions.
However, it makes `serializer.c` become way too large and messy, so
I move those tool functions out of `serializer.c` to `ser_help.c`,
and MAM relating functions to `ser_mam.c`.

This PR also fixed the error that non-tryte char in the `address`, and
`tag` field of a `send_transfer()` would cause SEGV fault.

Fixes #625
Closes #515
