+++ accelerator/core/serializer/serializer.c
@@ -7,14 +7,15 @@
  */
 
 #include "serializer.h"
+#include "ser_helper.h"
 #ifdef MQTT_ENABLE
 #include "connectivity/mqtt/mqtt_common.h"
 #endif
 #include "common/logger.h"
 #include "time.h"
-#define SERI_LOGGER "serializer"
 
-static logger_id_t logger_id;
+#define SERI_LOGGER "serializer"
+#define logger_id ser_logger_id
 
 void serializer_logger_init() { logger_id = logger_helper_enable(SERI_LOGGER, LOGGER_DEBUG, true); }
 
@@ -99,336 +100,6 @@ status_t db_identity_serialize(char** obj, db_identity_t* id_obj) {
 }
 #endif
 
-status_t string_utarray_to_json_array(UT_array const* const ut, cJSON* const json_root, char const* const obj_name) {
-  cJSON* array_obj = cJSON_CreateArray();
-  char** p = NULL;
-
-  if (!ut) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-
-  if (array_obj == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_JSON_CREATE));
-    return SC_SERIALIZER_JSON_CREATE;
-  }
-
-  cJSON_AddItemToObject(json_root, obj_name, array_obj);
-
-  while ((p = (char**)utarray_next(ut, p))) {
-    cJSON_AddItemToArray(array_obj, cJSON_CreateString(*p));
-  }
-  return SC_OK;
-}
-
-status_t json_string_array_to_string_utarray(cJSON const* const obj, char const* const obj_name, UT_array* const ut) {
-  char* str = NULL;
-
-  cJSON* json_item = cJSON_GetObjectItemCaseSensitive(obj, obj_name);
-  if (cJSON_IsArray(json_item)) {
-    cJSON* array_elt = NULL;
-    cJSON_ArrayForEach(array_elt, json_item) {
-      str = cJSON_GetStringValue(array_elt);
-      if (!str) {
-        ta_log_error("%s\n", "Encountered non-string array member");
-        return SC_SERIALIZER_JSON_PARSE;
-      }
-      utarray_push_back(ut, &str);
-    }
-  } else {
-    ta_log_error("%s\n", "Not an array");
-    return SC_SERIALIZER_JSON_PARSE;
-  }
-
-  return SC_OK;
-}
-
-static status_t ta_hash243_stack_to_json_array(hash243_stack_t stack, cJSON* json_root) {
-  size_t array_count = 0;
-  hash243_stack_entry_t* s_iter = NULL;
-  tryte_t trytes_out[NUM_TRYTES_HASH + 1];
-  size_t trits_count = 0;
-
-  array_count = hash243_stack_count(stack);
-  if (array_count > 0) {
-    LL_FOREACH(stack, s_iter) {
-      trits_count = flex_trits_to_trytes(trytes_out, NUM_TRYTES_HASH, s_iter->hash, NUM_TRITS_HASH, NUM_TRITS_HASH);
-      trytes_out[NUM_TRYTES_HASH] = '\0';
-      if (trits_count != 0) {
-        cJSON_AddItemToArray(json_root, cJSON_CreateString((const char*)trytes_out));
-      } else {
-        ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_INVALID_FLEX_TRITS));
-        return SC_CCLIENT_INVALID_FLEX_TRITS;
-      }
-    }
-  } else {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_NOT_FOUND));
-    return SC_CCLIENT_NOT_FOUND;
-  }
-  return SC_OK;
-}
-
-static status_t ta_json_array_to_hash243_queue(cJSON const* const obj, char const* const obj_name,
-                                               hash243_queue_t* queue) {
-  status_t ret_code = SC_OK;
-  flex_trit_t hash[FLEX_TRIT_SIZE_243] = {};
-  cJSON* json_item = cJSON_GetObjectItemCaseSensitive(obj, obj_name);
-  if (!json_item) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-
-  if (cJSON_IsArray(json_item)) {
-    cJSON* current_obj = NULL;
-    cJSON_ArrayForEach(current_obj, json_item) {
-      if (current_obj->valuestring != NULL) {
-        flex_trits_from_trytes(hash, NUM_TRITS_HASH, (tryte_t const*)current_obj->valuestring, NUM_TRYTES_HASH,
-                               NUM_TRYTES_HASH);
-        if (hash243_queue_push(queue, hash) != RC_OK) {
-          ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_JSON_PARSE));
-          return SC_SERIALIZER_JSON_PARSE;
-        }
-      }
-    }
-  } else {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_JSON_PARSE));
-    return SC_SERIALIZER_JSON_PARSE;
-  }
-  return ret_code;
-}
-
-static status_t ta_hash243_queue_to_json_array(hash243_queue_t queue, cJSON* const json_root) {
-  size_t array_count;
-  hash243_queue_entry_t* q_iter = NULL;
-
-  array_count = hash243_queue_count(queue);
-  if (array_count > 0) {
-    CDL_FOREACH(queue, q_iter) {
-      tryte_t trytes_out[NUM_TRYTES_HASH + 1];
-      size_t trits_count =
-          flex_trits_to_trytes(trytes_out, NUM_TRYTES_HASH, q_iter->hash, NUM_TRITS_HASH, NUM_TRITS_HASH);
-      trytes_out[NUM_TRYTES_HASH] = '\0';
-      if (trits_count != 0) {
-        cJSON_AddItemToArray(json_root, cJSON_CreateString((const char*)trytes_out));
-      } else {
-        ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_INVALID_FLEX_TRITS));
-        return SC_CCLIENT_INVALID_FLEX_TRITS;
-      }
-    }
-  } else {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_NOT_FOUND));
-    return SC_CCLIENT_NOT_FOUND;
-  }
-  return SC_OK;
-}
-
-static status_t ta_json_array_to_hash8019_array(cJSON const* const obj, char const* const obj_name,
-                                                hash8019_array_p array) {
-  status_t ret = SC_OK;
-  flex_trit_t hash[FLEX_TRIT_SIZE_8019] = {};
-  cJSON* json_item = cJSON_GetObjectItemCaseSensitive(obj, obj_name);
-  if (!cJSON_IsArray(json_item)) {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_JSON_PARSE));
-    return SC_CCLIENT_JSON_PARSE;
-  }
-
-  cJSON* current_obj = NULL;
-  cJSON_ArrayForEach(current_obj, json_item) {
-    if (current_obj->valuestring != NULL) {
-      if (strlen(current_obj->valuestring) != NUM_TRYTES_SERIALIZED_TRANSACTION) {
-        ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_INVALID_REQ));
-        return SC_SERIALIZER_INVALID_REQ;
-      }
-      flex_trits_from_trytes(hash, NUM_TRITS_SERIALIZED_TRANSACTION, (tryte_t const*)current_obj->valuestring,
-                             NUM_TRYTES_SERIALIZED_TRANSACTION, NUM_TRYTES_SERIALIZED_TRANSACTION);
-      hash_array_push(array, hash);
-    }
-  }
-  return ret;
-}
-
-status_t ta_hash8019_array_to_json_array(hash8019_array_p array, cJSON* const json_root, char const* const obj_name) {
-  size_t array_count = 0;
-  cJSON* array_obj = NULL;
-  tryte_t trytes_out[NUM_TRYTES_SERIALIZED_TRANSACTION + 1] = {};
-  size_t trits_count = 0;
-  flex_trit_t* elt = NULL;
-
-  array_count = hash_array_len(array);
-  if (array_count > 0) {
-    array_obj = cJSON_CreateArray();
-    if (array_obj == NULL) {
-      ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_JSON_CREATE));
-      return SC_SERIALIZER_JSON_CREATE;
-    }
-    cJSON_AddItemToObject(json_root, obj_name, array_obj);
-
-    HASH_ARRAY_FOREACH(array, elt) {
-      trits_count = flex_trits_to_trytes(trytes_out, NUM_TRYTES_SERIALIZED_TRANSACTION, elt,
-                                         NUM_TRITS_SERIALIZED_TRANSACTION, NUM_TRITS_SERIALIZED_TRANSACTION);
-      trytes_out[NUM_TRYTES_SERIALIZED_TRANSACTION] = '\0';
-      if (trits_count == 0) {
-        ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_FLEX_TRITS));
-        return SC_CCLIENT_FLEX_TRITS;
-      }
-      cJSON_AddItemToArray(array_obj, cJSON_CreateString((char const*)trytes_out));
-    }
-  }
-  return SC_OK;
-}
-
-static status_t ta_json_get_string(cJSON const* const json_obj, char const* const obj_name, char* const text,
-                                   const size_t size) {
-  status_t ret = SC_OK;
-  if (json_obj == NULL || obj_name == NULL || text == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-
-  cJSON* json_value = cJSON_GetObjectItemCaseSensitive(json_obj, obj_name);
-  if (json_value == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_JSON_KEY));
-    return SC_CCLIENT_JSON_KEY;
-  }
-
-  if (cJSON_IsString(json_value) && (json_value->valuestring != NULL)) {
-    strncpy(text, json_value->valuestring, size);
-  } else {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_JSON_PARSE));
-    return SC_CCLIENT_JSON_PARSE;
-  }
-
-  return ret;
-}
-
-status_t iota_transaction_to_json_object(iota_transaction_t const* const txn, cJSON** txn_json) {
-  if (txn == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_CCLIENT_NOT_FOUND));
-    return SC_CCLIENT_NOT_FOUND;
-  }
-  char msg_trytes[NUM_TRYTES_SIGNATURE + 1], hash_trytes[NUM_TRYTES_HASH + 1], tag_trytes[NUM_TRYTES_TAG + 1];
-
-  if (txn_json == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_JSON_CREATE));
-    return SC_SERIALIZER_JSON_CREATE;
-  }
-
-  // transaction hash
-  flex_trits_to_trytes((tryte_t*)hash_trytes, NUM_TRYTES_HASH, transaction_hash(txn), NUM_TRITS_HASH, NUM_TRITS_HASH);
-  hash_trytes[NUM_TRYTES_HASH] = '\0';
-  cJSON_AddStringToObject(*txn_json, "hash", hash_trytes);
-
-  // message
-  flex_trits_to_trytes((tryte_t*)msg_trytes, NUM_TRYTES_SIGNATURE, transaction_message(txn), NUM_TRITS_SIGNATURE,
-                       NUM_TRITS_SIGNATURE);
-  msg_trytes[NUM_TRYTES_SIGNATURE] = '\0';
-  cJSON_AddStringToObject(*txn_json, "signature_and_message_fragment", msg_trytes);
-
-  // address
-  flex_trits_to_trytes((tryte_t*)hash_trytes, NUM_TRYTES_HASH, transaction_address(txn), NUM_TRITS_HASH,
-                       NUM_TRITS_HASH);
-  hash_trytes[NUM_TRYTES_HASH] = '\0';
-  cJSON_AddStringToObject(*txn_json, "address", hash_trytes);
-  // value
-  cJSON_AddNumberToObject(*txn_json, "value", transaction_value(txn));
-  // obsolete tag
-  flex_trits_to_trytes((tryte_t*)tag_trytes, NUM_TRYTES_TAG, transaction_obsolete_tag(txn), NUM_TRITS_TAG,
-                       NUM_TRITS_TAG);
-  tag_trytes[NUM_TRYTES_TAG] = '\0';
-  cJSON_AddStringToObject(*txn_json, "obsolete_tag", tag_trytes);
-
-  // timestamp
-  cJSON_AddNumberToObject(*txn_json, "timestamp", transaction_timestamp(txn));
-
-  // current index
-  cJSON_AddNumberToObject(*txn_json, "current_index", transaction_current_index(txn));
-
-  // last index
-  cJSON_AddNumberToObject(*txn_json, "last_index", transaction_last_index(txn));
-
-  // bundle hash
-  flex_trits_to_trytes((tryte_t*)hash_trytes, NUM_TRYTES_HASH, transaction_bundle(txn), NUM_TRITS_HASH, NUM_TRITS_HASH);
-  hash_trytes[NUM_TRYTES_HASH] = '\0';
-  cJSON_AddStringToObject(*txn_json, "bundle_hash", hash_trytes);
-
-  // trunk transaction hash
-  flex_trits_to_trytes((tryte_t*)hash_trytes, NUM_TRYTES_HASH, transaction_trunk(txn), NUM_TRITS_HASH, NUM_TRITS_HASH);
-  hash_trytes[NUM_TRYTES_HASH] = '\0';
-  cJSON_AddStringToObject(*txn_json, "trunk_transaction_hash", hash_trytes);
-
-  // branch transaction hash
-  flex_trits_to_trytes((tryte_t*)hash_trytes, NUM_TRYTES_HASH, transaction_branch(txn), NUM_TRITS_HASH, NUM_TRITS_HASH);
-  hash_trytes[NUM_TRYTES_HASH] = '\0';
-  cJSON_AddStringToObject(*txn_json, "branch_transaction_hash", hash_trytes);
-
-  // tag
-  flex_trits_to_trytes((tryte_t*)tag_trytes, NUM_TRYTES_TAG, transaction_tag(txn), NUM_TRITS_TAG, NUM_TRITS_TAG);
-  tag_trytes[NUM_TRYTES_TAG] = '\0';
-  cJSON_AddStringToObject(*txn_json, "tag", tag_trytes);
-
-  // attachment timestamp
-  cJSON_AddNumberToObject(*txn_json, "attachment_timestamp", transaction_attachment_timestamp(txn));
-
-  // attachment lower timestamp
-  cJSON_AddNumberToObject(*txn_json, "attachment_timestamp_lower_bound", transaction_attachment_timestamp_lower(txn));
-
-  // attachment upper timestamp
-  cJSON_AddNumberToObject(*txn_json, "attachment_timestamp_upper_bound", transaction_attachment_timestamp_upper(txn));
-
-  // nonce
-  flex_trits_to_trytes((tryte_t*)tag_trytes, NUM_TRYTES_NONCE, transaction_nonce(txn), NUM_TRITS_NONCE,
-                       NUM_TRITS_NONCE);
-  tag_trytes[NUM_TRYTES_TAG] = '\0';
-  cJSON_AddStringToObject(*txn_json, "nonce", tag_trytes);
-
-  return SC_OK;
-}
-
-static status_t transaction_array_to_json_array(cJSON* json_root, const transaction_array_t* const txn_array) {
-  status_t ret = SC_OK;
-  iota_transaction_t* txn = NULL;
-
-  TX_OBJS_FOREACH(txn_array, txn) {
-    cJSON* txn_obj = cJSON_CreateObject();
-    if (!txn_obj) {
-      ret = SC_SERIALIZER_JSON_CREATE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-
-    ret = iota_transaction_to_json_object(txn, &txn_obj);
-    if (ret != SC_OK) {
-      cJSON_Delete(txn_obj);
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    cJSON_AddItemToArray(json_root, txn_obj);
-  }
-  return ret;
-}
-
-static status_t bundle_transaction_to_json_array(const bundle_transactions_t* const bundle, cJSON* json_root) {
-  status_t ret = SC_OK;
-  iota_transaction_t* txn = NULL;
-  BUNDLE_FOREACH(bundle, txn) {
-    cJSON* txn_obj = cJSON_CreateObject();
-    if (!txn_obj) {
-      ret = SC_SERIALIZER_JSON_CREATE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-
-    ret = iota_transaction_to_json_object(txn, &txn_obj);
-    if (ret != SC_OK) {
-      cJSON_Delete(txn_obj);
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    cJSON_AddItemToArray(json_root, txn_obj);
-  }
-  return ret;
-}
-
 status_t ta_generate_address_res_serialize(const ta_generate_address_res_t* const res, char** obj) {
   cJSON* json_root = cJSON_CreateArray();
   status_t ret = SC_OK;
@@ -486,7 +157,6 @@ status_t ta_send_transfer_req_deserialize(const char* const obj, ta_send_transfe
   cJSON* json_obj = cJSON_Parse(obj);
   cJSON* json_result = NULL;
   flex_trit_t tag_trits[NUM_TRITS_TAG], address_trits[NUM_TRITS_HASH];
-  int msg_len = 0, tag_len = 0;
   bool raw_message = true;
   status_t ret = SC_OK;
 
@@ -504,16 +174,37 @@ status_t ta_send_transfer_req_deserialize(const char* const obj, ta_send_transfe
     req->value = 0;
   }
 
+  json_result = cJSON_GetObjectItemCaseSensitive(json_obj, "address");
+  if (json_result != NULL && json_result->valuestring != NULL) {
+    int addr_len = strlen(json_result->valuestring);
+    if (!valid_tryte(json_result->valuestring, addr_len)) {
+      ret = SC_SERIALIZER_JSON_PARSE_NOT_TRYTE;
+      ta_log_error("%s\n", ta_error_to_string(ret));
+      goto done;
+    }
+
+    flex_trits_from_trytes(address_trits, NUM_TRITS_ADDRESS, (const tryte_t*)json_result->valuestring,
+                           NUM_TRYTES_ADDRESS, NUM_TRYTES_ADDRESS);
+  } else {
+    // 'address' does not exists, set to DEFAULT_ADDRESS
+    flex_trits_from_trytes(address_trits, NUM_TRITS_ADDRESS, (const tryte_t*)DEFAULT_ADDRESS, NUM_TRYTES_ADDRESS,
+                           NUM_TRYTES_ADDRESS);
+  }
+  ret = hash243_queue_push(&req->address, address_trits);
+  if (ret) {
+    ret = SC_CCLIENT_HASH;
+    ta_log_error("%s\n", ta_error_to_string(ret));
+    goto done;
+  }
+
   json_result = cJSON_GetObjectItemCaseSensitive(json_obj, "tag");
   if (json_result != NULL && json_result->valuestring != NULL) {
-    tag_len = strnlen(json_result->valuestring, NUM_TRYTES_TAG);
+    int tag_len = strlen(json_result->valuestring);
 
-    for (int i = 0; i < tag_len; i++) {
-      if (json_result->valuestring[i] & (unsigned)128) {
-        ret = SC_SERIALIZER_JSON_PARSE_ASCII;
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
+    if (!valid_tryte(json_result->valuestring, tag_len)) {
+      ret = SC_SERIALIZER_JSON_PARSE_NOT_TRYTE;
+      ta_log_error("%s\n", ta_error_to_string(ret));
+      goto done;
     }
 
     // If 'tag' is less than 27 trytes (NUM_TRYTES_TAG), expands it
@@ -548,13 +239,13 @@ status_t ta_send_transfer_req_deserialize(const char* const obj, ta_send_transfe
 
   json_result = cJSON_GetObjectItemCaseSensitive(json_obj, "message");
   if (json_result != NULL && json_result->valuestring != NULL) {
-    msg_len = strlen(json_result->valuestring);
+    int msg_len = strlen(json_result->valuestring);
 
     // In case the payload is unicode, the character whose ASCII code is beyond 128 result to an
     // error status_t code
     for (int i = 0; i < msg_len; i++) {
       if (json_result->valuestring[i] & 0x80) {
-        ret = SC_SERIALIZER_JSON_PARSE_ASCII;
+        ret = SC_SERIALIZER_JSON_PARSE_NOT_TRYTE;
         ta_log_error("%s\n", ta_error_to_string(ret));
         goto done;
       }
@@ -583,22 +274,6 @@ status_t ta_send_transfer_req_deserialize(const char* const obj, ta_send_transfe
     // 'message' does not exists, set to DEFAULT_MSG
     req->msg_len = DEFAULT_MSG_LEN * 3;
     flex_trits_from_trytes(req->message, req->msg_len, (const tryte_t*)DEFAULT_MSG, DEFAULT_MSG_LEN, DEFAULT_MSG_LEN);
-  }
-
-  json_result = cJSON_GetObjectItemCaseSensitive(json_obj, "address");
-  if (json_result != NULL && json_result->valuestring != NULL && (strnlen(json_result->valuestring, 81) == 81)) {
-    flex_trits_from_trytes(address_trits, NUM_TRITS_HASH, (const tryte_t*)json_result->valuestring, NUM_TRYTES_HASH,
-                           NUM_TRYTES_HASH);
-  } else {
-    // 'address' does not exists, set to DEFAULT_ADDRESS
-    flex_trits_from_trytes(address_trits, NUM_TRITS_HASH, (const tryte_t*)DEFAULT_ADDRESS, NUM_TRYTES_HASH,
-                           NUM_TRYTES_HASH);
-  }
-  ret = hash243_queue_push(&req->address, address_trits);
-  if (ret) {
-    ret = SC_CCLIENT_HASH;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
   }
 
 done:
@@ -639,7 +314,7 @@ status_t ta_send_trytes_res_serialize(const hash8019_array_p trytes, char** obj)
   status_t ret = SC_OK;
   cJSON* json_root = cJSON_CreateObject();
 
-  ret = ta_hash8019_array_to_json_array(trytes, json_root, "trytes");
+  ret = ta_hash8019_array_to_json_array(trytes, "trytes", json_root);
   if (ret != SC_OK) {
     goto done;
   }
@@ -686,7 +361,7 @@ status_t ta_find_transaction_object_single_res_serialize(transaction_array_t* re
     return SC_CCLIENT_JSON_CREATE;
   }
 
-  ret = iota_transaction_to_json_object(transaction_array_at(res, 0), &json_root);
+  ret = ta_iota_transaction_to_json_object(transaction_array_at(res, 0), &json_root);
   if (ret != SC_OK) {
     goto done;
   }
@@ -710,7 +385,7 @@ status_t ta_find_transaction_objects_res_serialize(const transaction_array_t* co
     return SC_CCLIENT_JSON_CREATE;
   }
 
-  ret = transaction_array_to_json_array(json_root, res);
+  ret = ta_transaction_array_to_json_array(res, json_root);
   if (ret != SC_OK) {
     goto done;
   }
@@ -758,109 +433,6 @@ done:
   return ret;
 }
 
-static status_t send_mam_message_mam_v1_req_deserialize(cJSON const* const json_obj, ta_send_mam_req_t* const req) {
-  cJSON *json_key = NULL, *json_value = NULL;
-  status_t ret = SC_OK;
-  send_mam_data_mam_v1_t* data = (send_mam_data_mam_v1_t*)req->data;
-  send_mam_key_mam_v1_t* key = (send_mam_key_mam_v1_t*)req->key;
-
-  ret = ta_json_get_string(json_obj, "x-api-key", req->service_token, SERVICE_TOKEN_LEN);
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  if (cJSON_HasObjectItem(json_obj, "key")) {
-    json_key = cJSON_GetObjectItem(json_obj, "key");
-    if (json_key == NULL) {
-      ret = SC_SERIALIZER_JSON_PARSE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-
-    if (cJSON_HasObjectItem(json_key, "psk")) {
-      ret = json_string_array_to_string_utarray(json_key, "psk", key->psk_array);
-      if (ret != SC_OK) {
-        ret = SC_SERIALIZER_JSON_PARSE;
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
-    }
-
-    if (cJSON_HasObjectItem(json_key, "ntru")) {
-      ret = json_string_array_to_string_utarray(json_key, "ntru", key->ntru_array);
-      if (ret != SC_OK) {
-        ret = SC_SERIALIZER_JSON_PARSE;
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
-    }
-  }
-
-  json_key = cJSON_GetObjectItem(json_obj, "data");
-  if (json_key == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  json_value = cJSON_GetObjectItemCaseSensitive(json_key, "seed");
-  if (json_value != NULL) {
-    size_t seed_size = strnlen(json_value->valuestring, NUM_TRYTES_ADDRESS);
-
-    if (seed_size != NUM_TRYTES_ADDRESS) {
-      ret = SC_SERIALIZER_INVALID_REQ;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-    data->seed = (tryte_t*)malloc(sizeof(tryte_t) * (NUM_TRYTES_ADDRESS + 1));
-    snprintf((char*)data->seed, seed_size + 1, "%s", json_value->valuestring);
-  }
-
-  if (cJSON_HasObjectItem(json_key, "chid")) {
-    json_value = cJSON_GetObjectItemCaseSensitive(json_key, "chid");
-    if (json_value != NULL) {
-      size_t chid_size = strnlen(json_value->valuestring, NUM_TRYTES_ADDRESS);
-
-      if (chid_size != NUM_TRYTES_ADDRESS) {
-        ret = SC_SERIALIZER_INVALID_REQ;
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
-      data->chid = (tryte_t*)malloc(sizeof(tryte_t) * (NUM_TRYTES_ADDRESS + 1));
-      snprintf((char*)data->chid, chid_size + 1, "%s", json_value->valuestring);
-    }
-  }
-
-  json_value = cJSON_GetObjectItemCaseSensitive(json_key, "message");
-  if (json_value != NULL) {
-    size_t payload_size = strlen(json_value->valuestring);
-    data->message = (char*)malloc((payload_size + 1) * sizeof(char));
-
-    // In case the payload is unicode, the character whose ASCII code is beyond 128 result to an
-    // error status_t code
-    for (size_t i = 0; i < payload_size; i++) {
-      if (json_value->valuestring[i] & 0x80) {
-        ret = SC_SERIALIZER_JSON_PARSE_ASCII;
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
-    }
-    memcpy(data->message, json_value->valuestring, payload_size + 1);
-  } else {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    ret = SC_SERIALIZER_NULL;
-  }
-
-  json_value = cJSON_GetObjectItemCaseSensitive(json_key, "ch_mss_depth");
-  if ((json_value != NULL) && cJSON_IsNumber(json_value)) {
-    data->ch_mss_depth = json_value->valueint;
-  }
-
-done:
-  return ret;
-}
-
 status_t ta_send_transfer_res_serialize(ta_send_transfer_res_t* res, char** obj) {
   status_t ret = SC_OK;
   cJSON* json_root = cJSON_CreateObject();
@@ -874,7 +446,7 @@ status_t ta_send_transfer_res_serialize(ta_send_transfer_res_t* res, char** obj)
     cJSON_AddStringToObject(json_root, "uuid", res->uuid);
     cJSON_AddStringToObject(json_root, "address", (char*)res->address);
   } else {
-    ret = iota_transaction_to_json_object(transaction_array_at(res->txn_array, 0), &json_root);
+    ret = ta_iota_transaction_to_json_object(transaction_array_at(res->txn_array, 0), &json_root);
     if (ret != SC_OK) {
       ta_log_error("%s\n", error_2_string(ret));
       goto done;
@@ -893,371 +465,6 @@ status_t ta_send_transfer_res_serialize(ta_send_transfer_res_t* res, char** obj)
 
 done:
   cJSON_Delete(json_root);
-  return ret;
-}
-
-status_t send_mam_req_deserialize(const char* const obj, ta_send_mam_req_t* const req) {
-  if (obj == NULL || req == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-  status_t ret = SC_OK;
-  cJSON* json_obj = cJSON_Parse(obj);
-
-  const int protocol_len = 20;
-  char protocol_str[protocol_len];
-
-  if (json_obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  ret = ta_json_get_string(json_obj, "protocol", protocol_str, protocol_len);
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-  if (!strncmp(protocol_str, "MAM_V1", strlen("MAM_V1"))) {
-    req->protocol = MAM_V1;
-  }
-
-  send_mam_req_v1_init(req);
-  ret = send_mam_message_mam_v1_req_deserialize(json_obj, req);
-  if (ret) {
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-done:
-  cJSON_Delete(json_obj);
-  return ret;
-}
-
-static status_t recv_mam_message_mam_v1_req_deserialize(cJSON const* const json_obj, ta_recv_mam_req_t* const req) {
-  cJSON *json_key = NULL, *json_value = NULL;
-  status_t ret = SC_OK;
-  char *bundle_hash = NULL, *chid = NULL, *msg_id = NULL, *psk = NULL, *ntru = NULL;
-
-  if (cJSON_HasObjectItem(json_obj, "key")) {
-    json_value = cJSON_GetObjectItemCaseSensitive(json_obj, "key");
-    if (json_value == NULL) {
-      ret = SC_CCLIENT_JSON_KEY;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    if (cJSON_IsString(json_value) && (json_value->valuestring != NULL)) {
-      if (strlen(json_value->valuestring) == NUM_TRYTES_MAM_PSK_KEY_SIZE) {
-        psk = (char*)malloc(sizeof(char) * (NUM_TRYTES_MAM_PSK_KEY_SIZE + 1));
-        strncpy(psk, json_value->valuestring, sizeof(char) * (NUM_TRYTES_MAM_PSK_KEY_SIZE + 1));
-      } else if (strlen(json_value->valuestring) == NUM_TRYTES_MAM_NTRU_PK_SIZE) {
-        ntru = (char*)malloc(sizeof(char) * (NUM_TRYTES_MAM_NTRU_PK_SIZE + 1));
-        strncpy(ntru, json_value->valuestring, sizeof(char) * (NUM_TRYTES_MAM_NTRU_PK_SIZE + 1));
-      }
-    } else {
-      ret = SC_CCLIENT_JSON_PARSE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-
-    ret = recv_mam_set_mam_v1_key(req, (tryte_t*)psk, (tryte_t*)ntru);
-    if (ret != SC_OK) {
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-  }
-
-  json_key = cJSON_GetObjectItem(json_obj, "data_id");
-  if (json_key == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  if (cJSON_HasObjectItem(json_key, "bundle_hash")) {
-    json_value = cJSON_GetObjectItemCaseSensitive(json_key, "bundle_hash");
-    if (json_value == NULL) {
-      ret = SC_CCLIENT_JSON_KEY;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    if (cJSON_IsString(json_value) && (json_value->valuestring != NULL) &&
-        (strlen(json_value->valuestring) == NUM_TRYTES_HASH)) {
-      bundle_hash = (char*)malloc(sizeof(char) * (NUM_TRYTES_HASH + 1));
-      strncpy(bundle_hash, json_value->valuestring, sizeof(char) * (NUM_TRYTES_HASH + 1));
-    } else {
-      ret = SC_CCLIENT_JSON_PARSE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    goto set_data_id;
-  }
-
-  if (cJSON_HasObjectItem(json_key, "chid")) {
-    json_value = cJSON_GetObjectItemCaseSensitive(json_key, "chid");
-    if (json_value == NULL) {
-      ret = SC_CCLIENT_JSON_KEY;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    if (cJSON_IsString(json_value) &&
-        (json_value->valuestring != NULL && (strlen(json_value->valuestring) == NUM_TRYTES_HASH))) {
-      chid = (char*)malloc(sizeof(char) * (NUM_TRYTES_HASH + 1));
-      strncpy(chid, json_value->valuestring, sizeof(char) * (NUM_TRYTES_HASH + 1));
-    } else {
-      ret = SC_CCLIENT_JSON_PARSE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-  }
-
-  if (cJSON_HasObjectItem(json_key, "msg_id")) {
-    json_value = cJSON_GetObjectItemCaseSensitive(json_key, "msg_id");
-    if (json_value == NULL) {
-      ret = SC_CCLIENT_JSON_KEY;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-    if (cJSON_IsString(json_value) && (json_value->valuestring != NULL) &&
-        (strlen(json_value->valuestring) == NUM_TRYTES_MAM_MSG_ID)) {
-      msg_id = (char*)malloc(sizeof(char) * (NUM_TRYTES_MAM_MSG_ID + 1));
-      strncpy(msg_id, json_value->valuestring, sizeof(char) * (NUM_TRYTES_MAM_MSG_ID + 1));
-    } else {
-      ret = SC_CCLIENT_JSON_PARSE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      return ret;
-    }
-  }
-
-set_data_id:
-  ret = recv_mam_set_mam_v1_data_id(req, bundle_hash, chid, msg_id);
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-done:
-  free(psk);
-  free(ntru);
-  free(bundle_hash);
-  free(chid);
-  free(msg_id);
-  return ret;
-}
-
-status_t recv_mam_message_req_deserialize(const char* const obj, ta_recv_mam_req_t* const req) {
-  if (obj == NULL || req == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-  status_t ret = SC_OK;
-  cJSON* json_obj = cJSON_Parse(obj);
-
-  const int protocol_len = 20;
-  char protocol_str[protocol_len];
-
-  if (json_obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  ret = ta_json_get_string(json_obj, "protocol", protocol_str, protocol_len);
-  if (ret != SC_OK) {
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-  if (!strncmp(protocol_str, "MAM_V1", strlen("MAM_V1"))) {
-    req->protocol = MAM_V1;
-  }
-
-  switch (req->protocol) {
-    case MAM_V1:
-      ret = recv_mam_message_mam_v1_req_deserialize(json_obj, req);
-      if (ret) {
-        ta_log_error("%s\n", ta_error_to_string(ret));
-        goto done;
-      }
-      break;
-
-    default:
-      break;
-  }
-
-done:
-  cJSON_Delete(json_obj);
-  return ret;
-}
-
-status_t recv_mam_res_deserialize(const char* const obj, ta_recv_mam_res_t* const res) {
-  status_t ret = SC_OK;
-  if (!obj || !res) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-  cJSON* json_obj = cJSON_Parse(obj);
-
-  if (json_obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  ret = json_string_array_to_string_utarray(json_obj, "payload", res->payload_array);
-  if (ret) {
-    ret = SC_CCLIENT_JSON_KEY;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  if (cJSON_HasObjectItem(json_obj, "chid1")) {
-    ret = ta_json_get_string(json_obj, "chid1", res->chid1, NUM_TRYTES_ADDRESS);
-    if (ret) {
-      ret = SC_CCLIENT_JSON_KEY;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-    }
-  }
-
-done:
-  cJSON_Delete(json_obj);
-  return ret;
-}
-
-status_t recv_mam_message_res_serialize(ta_recv_mam_res_t* const res, char** obj) {
-  status_t ret = SC_OK;
-  if (!res || !obj) {
-    ret = SC_SERIALIZER_NULL;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    return ret;
-  }
-
-  cJSON* json_root = cJSON_CreateObject();
-  if (json_root == NULL) {
-    ret = SC_SERIALIZER_JSON_CREATE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  ret = string_utarray_to_json_array(res->payload_array, json_root, "payload");
-  if (ret) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  if (res->chid1[0]) {
-    cJSON_AddStringToObject(json_root, "chid1", res->chid1);
-  }
-
-  *obj = cJSON_PrintUnformatted(json_root);
-  if (*obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-done:
-  cJSON_Delete(json_root);
-  return ret;
-}
-
-status_t send_mam_res_serialize(const ta_send_mam_res_t* const res, char** obj) {
-  status_t ret = SC_OK;
-  if (!res || !obj) {
-    ret = SC_SERIALIZER_NULL;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    return ret;
-  }
-
-  cJSON* json_root = cJSON_CreateObject();
-  if (json_root == NULL) {
-    ret = SC_SERIALIZER_JSON_CREATE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  cJSON_AddStringToObject(json_root, "bundle_hash", res->bundle_hash);
-
-  cJSON_AddStringToObject(json_root, "chid", res->chid);
-
-  cJSON_AddStringToObject(json_root, "msg_id", res->msg_id);
-
-  if (res->announcement_bundle_hash[0]) {
-    cJSON_AddStringToObject(json_root, "announcement_bundle_hash", res->announcement_bundle_hash);
-  }
-
-  if (res->chid1[0]) {
-    cJSON_AddStringToObject(json_root, "chid1", res->chid1);
-  }
-
-  *obj = cJSON_PrintUnformatted(json_root);
-  if (*obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-done:
-  cJSON_Delete(json_root);
-  return ret;
-}
-
-status_t send_mam_res_deserialize(const char* const obj, ta_send_mam_res_t* const res) {
-  if (obj == NULL) {
-    ta_log_error("%s\n", ta_error_to_string(SC_SERIALIZER_NULL));
-    return SC_SERIALIZER_NULL;
-  }
-  cJSON* json_obj = cJSON_Parse(obj);
-  status_t ret = SC_OK;
-  tryte_t addr[NUM_TRYTES_ADDRESS + 1], msg_id[MAM_MSG_ID_SIZE];
-
-  if (json_obj == NULL) {
-    ret = SC_SERIALIZER_JSON_PARSE;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-
-  if (ta_json_get_string(json_obj, "chid", (char*)addr, NUM_TRYTES_ADDRESS) != SC_OK) {
-    ret = SC_SERIALIZER_NULL;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-  send_mam_res_set_channel_id(res, addr);
-
-  if (ta_json_get_string(json_obj, "msg_id", (char*)msg_id, NUM_TRYTES_MAM_MSG_ID) != SC_OK) {
-    ret = SC_SERIALIZER_NULL;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-  send_mam_res_set_msg_id(res, msg_id);
-
-  if (ta_json_get_string(json_obj, "bundle_hash", (char*)addr, NUM_TRYTES_ADDRESS) != SC_OK) {
-    ret = SC_SERIALIZER_NULL;
-    ta_log_error("%s\n", ta_error_to_string(ret));
-    goto done;
-  }
-  send_mam_res_set_bundle_hash(res, addr);
-
-  if (cJSON_HasObjectItem(json_obj, "chid1")) {
-    if (ta_json_get_string(json_obj, "chid1", (char*)addr, NUM_TRYTES_ADDRESS) != SC_OK) {
-      ret = SC_SERIALIZER_NULL;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-    send_mam_res_set_chid1(res, addr);
-  }
-
-  if (cJSON_HasObjectItem(json_obj, "announcement_bundle_hash")) {
-    if (ta_json_get_string(json_obj, "announcement_bundle_hash", (char*)addr, NUM_TRYTES_ADDRESS) != SC_OK) {
-      ret = SC_SERIALIZER_NULL;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-    send_mam_res_set_announce_bundle_hash(res, addr);
-  }
-
-done:
-  cJSON_Delete(json_obj);
   return ret;
 }
 
@@ -1447,19 +654,11 @@ status_t fetch_txn_with_uuid_res_serialize(const ta_fetch_txn_with_uuid_res_t* c
   }
 
   if (res->status == SENT) {
-    cJSON* json_array = cJSON_CreateArray();
-    if (json_array == NULL) {
-      ret = SC_SERIALIZER_JSON_CREATE;
-      ta_log_error("%s\n", ta_error_to_string(ret));
-      goto done;
-    }
-
-    ret = bundle_transaction_to_json_array(res->bundle, json_array);
+    ret = ta_bundle_transaction_to_json_array(res->bundle, "bundle", json_root);
     if (ret) {
       ta_log_error("%s\n", ta_error_to_string(ret));
       goto done;
     }
-    cJSON_AddItemToObject(json_root, "bundle", json_array);
   } else if (res->status == UNSENT) {
     cJSON_AddStringToObject(json_root, "status", "unsent");
   } else if (res->status == NOT_EXIST) {
