+++ common/ta_errors.h
@@ -110,8 +110,8 @@ typedef enum {
   /**< NULL object in serializer */
   SC_SERIALIZER_JSON_PARSE = 0x03 | SC_MODULE_SERIALIZER | SC_SEVERITY_FATAL,
   /**< Failed to parse JSON object in serializer */
-  SC_SERIALIZER_JSON_PARSE_ASCII = 0x04 | SC_MODULE_SERIALIZER | SC_SEVERITY_FATAL,
-  /**< Unicode value in JSON */
+  SC_SERIALIZER_JSON_PARSE_NOT_TRYTE = 0x04 | SC_MODULE_SERIALIZER | SC_SEVERITY_FATAL,
+  /**< Non-tryte value in JSON */
   SC_SERIALIZER_INVALID_REQ = 0x05 | SC_MODULE_SERIALIZER | SC_SEVERITY_FATAL,
   /**< Invald request value in JSON */
   SC_SERIALIZER_MESSAGE_OVERRUN = 0x06 | SC_MODULE_SERIALIZER | SC_SEVERITY_FATAL,
